Put log streaming into one file per class
This is more in line with expected ruby conventions.
